Safari-609.1.11.1/LayoutTests/media/modern-media-controls/fullscreen-button/fullscreen-button-expected.txt

Testing the FullscreenButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS enterFullscreenButton.element.localName is "button"
PASS enterFullscreenButton.element.classList.contains("fullscreen") is true
PASS enterFullscreenButton.iconName is Icons.EnterFullscreen
PASS exitFullscreenButton.iconName is Icons.ExitFullscreen
PASS enterFullscreenButton.image.element.style.webkitMaskImage.includes("macOS/EnterFullscreen.svg") is true
PASS iOSEnterFullscreenButton.image.element.style.webkitMaskImage.includes("iOS/EnterFullscreen.svg") is true
PASS exitFullscreenButton.image.element.style.webkitMaskImage.includes("macOS/ExitFullscreen.svg") is true
PASS successfullyParsed is true

TEST COMPLETE
